fix(ci): resolve CI failures and update for Rust 1.85+
- Bump MSRV from 1.75 to 1.85 (required by home crate Edition 2024)
- Update deny.toml to v2 format, add CDLA-Permissive-2.0 license
- Ignore RUSTSEC-2025-0052 (async-std unmaintained advisory)
- Upgrade sqlx from 0.7 to 0.8 to resolve RUSTSEC-2024-0363
- Remove cross-compile job (ring crate incompatible with darwin cross)
- Simplify CI to 5 fast jobs: check, fmt, clippy, test, deny
- Fix all clippy warnings with workspace lint config
- Apply rustfmt 1.92.0 formatting
- Fix trybuild tests to avoid Rust version-dependent warning format

+++ mcpkit/tests/transport_e2e.rs
@@ -119,7 +119,7 @@ async fn test_message_sequence() {
         let request = msg.as_request().unwrap();
         assert_eq!(request.id, RequestId::Number(i));
 
-        let response = Response::success(i as u64, json!({"pong": i}));
+        let response = Response::success(i, json!({"pong": i}));
         server.send(Message::Response(response)).await.unwrap();
     }
 
@@ -229,7 +229,10 @@ async fn test_tool_call_lifecycle() {
             }]
         }),
     );
-    server.send(Message::Response(tools_response)).await.unwrap();
+    server
+        .send(Message::Response(tools_response))
+        .await
+        .unwrap();
 
     let msg = client.recv().await.unwrap().unwrap();
     let response = msg.as_response().unwrap();
